+++ apps/api/src/kick_logs/domain/entities/__init__.py
@@ -4,5 +4,14 @@ from kick_logs.domain.entities.emote import Emote
 from kick_logs.domain.entities.raw_kick_event import RawKickEvent
 from kick_logs.domain.entities.sender import Sender
 from kick_logs.domain.entities.user import User
+from kick_logs.domain.entities.worker_heartbeat import WorkerHeartbeat
 
-__all__ = ["Channel", "ChatMessage", "Emote", "RawKickEvent", "Sender", "User"]
+__all__ = [
+    "Channel",
+    "ChatMessage",
+    "Emote",
+    "RawKickEvent",
+    "Sender",
+    "User",
+    "WorkerHeartbeat",
+]